Move FiatAmount to a util
